+++ docs/changelog.md
@@ -108,12 +108,15 @@
 - new trait: `Word`.
 
 ### data::codec
-- new macro: `set!`.
 - new submodules: `bin`, `detect`, `pack`, `symbol`.
 - move `hash::check` to `integrity`.
 - move `bit` inside `bin`.
 - move `schema` inside `data::value`.
 - move `deser` inside `data::value::schema`
+
+### data::codec::bin
+- new macro: `set!`.
+- new types: `BinTag4`.
 - recreate `bitfield!` making it much more lean and improved.
 - remove the `bitfield` example.